Creation of Ditto management clasess
It has been created the dittoManager package, which contains two files at the moment: DittoThing.kt is the one that contains all the needed data structures for managing Ditto Things in Kotlin through json. The other one is the DittoConnectionManager class which contains all the necessary methods to connect to Ditto. At the moment those methods are PUT for creating/modifying, DELETE and GET. On the other hand the PeriodicDittoService has been refactored in order to contain only methods related with the periodic service