PeriodicDittoService.kt debugging and first rating method
The class PeriodicDittoService has been debugged and some methods have been changed

On the other hand, a first implementation of the exercise rating method has been done. To rate the session it takes into account the heart rates during the workout, the time spent in each HR zone and the work-to-rest ratio. It will also use the world record in each distance to make the rating not relative to each athlete